+++ Ropu.Shared/CallManagement/ICallManagementServerMessageHandler.cs
@@ -4,11 +4,11 @@ namespace Ropu.Shared.CallManagement
 {
     public interface ICallManagementServerMessageHandler
     {
-        void HandleRegisterMediaController(IPAddress from, uint requestId, ushort controlPort, IPEndPoint mediaEndpoint);
-        void HandleRegisterFloorController(IPAddress from, uint requestId, ushort controlPort, IPEndPoint floorControlEndpoint);
-        void HandleGetGroupsFileRequest(IPEndPoint from, uint requestId);
-        void HandleGetGroupFileRequest(IPEndPoint from, uint requestId, ushort groupId);
-        void HandleFilePartRequest(IPEndPoint from, uint requestId, ushort fileId, ushort partNumber);
-        void HandleCompleteFileTransfer(IPEndPoint from, uint requestId, ushort fileId);
+        void HandleRegisterMediaController(IPAddress from, ushort requestId, ushort controlPort, IPEndPoint mediaEndpoint);
+        void HandleRegisterFloorController(IPAddress from, ushort requestId, ushort controlPort, IPEndPoint floorControlEndpoint);
+        void HandleGetGroupsFileRequest(IPEndPoint from, ushort requestId);
+        void HandleGetGroupFileRequest(IPEndPoint from, ushort requestId, ushort groupId);
+        void HandleFilePartRequest(IPEndPoint from, ushort requestId, ushort fileId, ushort partNumber);
+        void HandleCompleteFileTransfer(IPEndPoint from, ushort requestId, ushort fileId);
     }
 }
